feat: move task options into side panel
Made-with: Cursor
