+++ src/components/OptionEditor.tsx
@@ -180,7 +180,7 @@ function OptionDescription({
       {/* 内容 */}
       {resolved.html && (
         <div
-          className="text-xs text-text-secondary [&_p]:my-0.5 [&_a]:text-accent [&_a]:hover:underline"
+          className="text-[11px] text-text-secondary [&_p]:my-0.5 [&_a]:text-accent [&_a]:hover:underline"
           dangerouslySetInnerHTML={{ __html: resolved.html }}
         />
       )}
